compute a rotation matrix using Euler angles, and include scaling
to avoid the mMultiply function, which I cannot compile due to the

error: unable to find a register to spill in class 'POINTER_REGS'

compiler error (is said to be a gcc bug).

Fixed Point Arithmetic is not yet clear to me, therefore the
not-very-general formulation of the matrix addition.

Added simple object glider.h, created from a quick tinkercad drawing,
exported to stl there and converted with stl3h.py .

The translation is a function of the Arduino Gamer Potis, to test it.